chore: checks added for non existing files

+++ scripts/install_parts.sh
@@ -6,12 +6,25 @@ cp rootro /usr/local/bin/
 ln -s rootro /usr/local/bin/rootrw 
 cp syncoverlayfs.service /lib/systemd/system/
 
-mv /etc/wpa_supplicant/wpa_supplicant.conf /var/local/
+# If wpa_supplicant.conf exists, move it to /var/local else create wpa_supplicant.conf at /var/local
+if [ -f /etc/wpa_supplicant/wpa_supplicant.conf ]; then
+    echo "wpa_supplicant.conf exists, moving to /var/local/"
+    mv /etc/wpa_supplicant/wpa_supplicant.conf /var/local/
+else
+    echo "wpa_supplicant.conf does not exist, creating at /var/local/"
+    touch /var/local/wpa_supplicant.conf
+fi
+
 ln -s /var/local/wpa_supplicant.conf /etc/wpa_supplicant/wpa_supplicant.conf
 mv /etc/resolv.conf /var/local/
 ln -s /var/local/resolv.conf /etc/resolv.conf
 touch /var/local/resolv.conf.bak
-rm /etc/resolv.conf.bak
+
+if [ -f /etc/resolv.conf.bak ]; then
+    echo "resolv.conf.bak exists, moving to /var/local/"
+    mv /etc/resolv.conf.bak /var/local/
+fi
+
 ln -s /var/local/resolv.conf.bak /etc/resolv.conf.bak
 mv /etc/fake-hwclock.data /var/local/
 ln -s /var/local/fake-hwclock.data /etc/fake-hwclock.data
